backend/requirements.txt
Django==3.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-telegram-bot==20.7
python-dotenv==1.0.0
Pillow==10.1.0
mysqlclient==2.2.0
requests==2.31.0
